refactor : channel, chatting interface 분리
